+++ frontend/src/app/dm/page.tsx
@@ -16,8 +16,18 @@ export default function DmListPage() {
 
   useEffect(() => {
     setDmUsers([
-      { id: 1, name: "さくら", lastMessage: "こんにちは！", avatarUrl: "/avatar1.jpg" },
-      { id: 2, name: "たけし", lastMessage: "またね！", avatarUrl: "/avatar2.jpg" },
+      {
+        id: 1,
+        name: "さくら",
+        lastMessage: "こんにちは！",
+        avatarUrl: "/avatar1.jpg",
+      },
+      {
+        id: 2,
+        name: "たけし",
+        lastMessage: "またね！",
+        avatarUrl: "/avatar2.jpg",
+      },
     ]);
   }, []);
 
@@ -30,8 +40,15 @@ export default function DmListPage() {
       <h1 className="title">DM一覧</h1>
       <ul className="dm-list">
         {dmUsers.map((user) => (
-          <li key={user.id} className="dm-item" onClick={() => handleClick(user.id)}>
-            <img src={user.avatarUrl || "/default-avatar.png"} className="avatar" />
+          <li
+            key={user.id}
+            className="dm-item"
+            onClick={() => handleClick(user.id)}
+          >
+            <img
+              src={user.avatarUrl || "/default-avatar.png"}
+              className="avatar"
+            />
             <div className="dm-content">
               <p className="dm-name">{user.name}</p>
               <p className="dm-last">{user.lastMessage}</p>
